@@ -19,9 +19,9 @@
   },
   "devDependencies": {
     "@sveltejs/adapter-vercel": "^4.0.5",
-    "@sveltejs/enhanced-img": "^0.3.0",
-    "@sveltejs/kit": "^2.5.18",
-    "@sveltejs/vite-plugin-svelte": "^3.1.1",
+    "@sveltejs/enhanced-img": "^0.3.9",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/forms": "^0.5.7",
     "@tailwindcss/typography": "^0.5.13",
     "@typescript-eslint/eslint-plugin": "^7.17.0",
@@ -30,21 +30,21 @@
     "cssnano": "^6.1.2",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.43.0",
+    "eslint-plugin-svelte": "^2.45.1",
     "imagetools-core": "^7.0.0",
     "mapbox-gl": "^3.2.0",
     "postcss": "^8.4.38",
     "postcss-load-config": "^5.0.3",
     "prettier": "3.2.5",
     "prettier-plugin-svelte": "^3.2.6",
-    "svelte": "^4.2.18",
-    "svelte-check": "^3.8.4",
+    "svelte": "^5.33.19",
+    "svelte-check": "^4.0.0",
     "svelte-preprocess": "^6.0.2",
     "tailwind-scrollbar": "^3.1.0",
     "tailwindcss": "^3.4.6",
     "tslib": "^2.6.2",
     "typescript": "^5.5.4",
-    "vite": "^5.2.7",
+    "vite": "^5.4.4",
     "vite-imagetools": "^7.0.4"
   },
   "optionalDependencies": {
