@@ -1,17 +1,28 @@
 <script lang="ts">
-  export let buttonSize: 'Standard' | 'Small'
-  export let buttonMargin: 'Standard' | 'None' = 'Standard'
-  export let reference = ''
-  export let label = ''
-  export let target = ''
-  export let clicked: () => void = () => {}
+  interface Props {
+    buttonSize: 'Standard' | 'Small';
+    buttonMargin?: 'Standard' | 'None';
+    reference?: string;
+    label?: string;
+    target?: string;
+    clicked?: () => void;
+  }
+
+  let {
+    buttonSize,
+    buttonMargin = 'Standard',
+    reference = '',
+    label = '',
+    target = '',
+    clicked = () => {}
+  }: Props = $props();
 </script>
 
 <a
   href={reference}
   type="button"
   target={target ?? ''}
-  on:click={clicked}
+  onclick={clicked}
   class="inline-flex md:inline-block justify-center items-center w-full sm:w-auto
   {buttonMargin === 'Standard' ? 'my-4' : 'my-0'} px-6 py-3
   {buttonSize === 'Standard' ? 'md:px-6 md:py-3' : 'md:px-3 md:py-1'}
