@@ -1,26 +1,36 @@
 <script lang="ts">
-  export let title: string
-  export let description: string
-  export let path: string
-  export let close: () => void
-  export let inline = false
+  interface Props {
+    title: string;
+    description: string;
+    path: string;
+    close: () => void;
+    inline?: boolean;
+  }
+
+  let {
+    title,
+    description,
+    path,
+    close,
+    inline = false
+  }: Props = $props();
 </script>
 
 <a
   href={path}
-  on:click={close}
+  onclick={close}
   class="relative {inline
     ? 'inline-link'
     : 'lg:dialog-link inline-link'} block group hover:bg-blue-triarc rounded hover:bg-opacity-10"
 >
   <div class="font-semibold text-gray-900 group-hover:text-blue-triarc">
     {title}
-    <span class="absolute inset-0" />
+    <span class="absolute inset-0"></span>
   </div>
   <p class="text-gray-600">{description}</p>
 </a>
 
-<!-- svelte-ignore css-unused-selector -->
+<!-- svelte-ignore css_unused_selector -->
 <style lang="postcss">
   /*noinspection CssUnusedSymbol*/
   .dialog-link {
