@@ -1,9 +1,18 @@
 <script lang="ts">
   import type { Picture } from 'imagetools-core'
-  export let imgClass: string = ''
-  export let alt = 'No alt available'
-  export let image: Picture
-  export let loading: 'eager' | 'lazy' = 'lazy'
+  interface Props {
+    imgClass?: string;
+    alt?: string;
+    image: Picture;
+    loading?: 'eager' | 'lazy';
+  }
+
+  let {
+    imgClass = '',
+    alt = 'No alt available',
+    image,
+    loading = 'lazy'
+  }: Props = $props();
 </script>
 
 <enhanced:img src={image} {alt} class={imgClass} {loading} />
