@@ -2,11 +2,15 @@
   // Not in one component because having an if for the video slowed down the display an insane amount
   import 'plyr/dist/plyr.css'
 
-  export let content: VideoContent
 
   import { onMount } from 'svelte'
   import Plyr from 'plyr'
   import type { VideoContent } from '$lib/components/TypeDefinitions'
+  interface Props {
+    content: VideoContent;
+  }
+
+  let { content }: Props = $props();
 
   onMount(async () => {
     new Plyr(`#${content.videoId}-player`, {
@@ -15,7 +19,7 @@
   })
 </script>
 
-<!-- svelte-ignore a11y-media-has-caption -->
+<!-- svelte-ignore a11y_media_has_caption -->
 <video
   id={`${content.videoId}-player`}
   playsinline
