@@ -11,7 +11,7 @@
 
   let listedJob = 'Senior Software Engineer'
 
-  let block = {
+  let block = $state({
     title: '<span class="text-sm">wir suchen</span><br/>Senior Software Engineer',
     content:
       'Als Senior begleitest du Projekte vom Requirements-Engineering bis in den Betrieb. Du bist der Vermittler zwischen dem Kunden und deinen Entwickler-KollegInnen. Und du machst das nicht zum ersten Mal. Du kannst im Team unterstützen, wo es an was fehlt.',
@@ -22,7 +22,7 @@
       'Erfahrungen mit Elasticsearch / Postgres / MongoDB',
       'Erfahrungen Docker / Kubernetes / Gitlab CI',
     ],
-  }
+  })
 </script>
 
 <svelte:head>
