@@ -108,7 +108,7 @@
             </blockquote>
           </div>
           <div class="flex">
-            <div class="bg-gray-300 w-6 h-auto rounded-full my-1" />
+            <div class="bg-gray-300 w-6 h-auto rounded-full my-1"></div>
             <p class="pl-4 text-gray-900 leading-7 italic">
               Preis- und Zeit-optimiertes Bestellwesen. Eine mobile Applikation zum effizienten Bestellen von benötigten
               Artikeln ohne sich um Details, wie bei welchem Lieferanten bestellt werden soll, kümmern zu müssen. Vom
@@ -162,7 +162,7 @@
             </blockquote>
           </div>
           <div class="flex">
-            <div class="bg-gray-300 w-6 h-auto rounded-full my-1" />
+            <div class="bg-gray-300 w-6 h-auto rounded-full my-1"></div>
             <p class="pl-4 text-gray-900 leading-7 italic">
               Die digitale Kundenbegleiterin, gelöst als mobil einsetzbare Applikation zur Erstellung der Anamnese,
               Analyse bis hin zur fallbezogenen und interdisziplinären Kommunikation. Damit stellt OrthoClient eine
