@@ -6,14 +6,14 @@
   import joinTheTeam from '$lib/assets/img/intro/content_team.svg'
   import heroImage from '$lib/assets/hero/Team.jpg?width=300;600;1000;2000&format=webp&enhanced'
 
-  let content = {
+  let content = $state({
     title: 'Werde Teil des Teams!',
     content:
       'Zusammen können wir alles erreichen. <br/><br/>' +
       'Werde Teil des Teams und bewirb dich, wenn du Leidenschaft, Elan und Tatendrang hast und du dich mit unseren Werten identifizieren kannst. ',
     image: { src: joinTheTeam, alt: 'Triarc Career' },
     link: { href: '/jobs', text: 'Unsere Jobs' },
-  }
+  })
 </script>
 
 <svelte:head>
